Make KeywordInfo generic
